refactor: 사용하지 않는 메서드 삭제
