+++ src/main/java/chess/domain/piece/None.java
@@ -24,11 +24,6 @@ public class None implements ChessPiece {
     }
 
     @Override
-    public void move() {
-        throw new IllegalStateException("움직일 말이 존재하지 않습니다.");
-    }
-
-    @Override
     public String name() {
         return "-";
     }
